Added loading of CKB from json

+++ Source/CiF/Private/CiFSocialNetwork.cpp
@@ -17,7 +17,7 @@ void UCiFSocialNetwork::init(const ESocialNetworkType networkType, const uint8 n
 	setAllArrayElements(maxVal / 2);
 }
 
-void UCiFSocialNetwork::setWeight(const int32 c1, const int32 c2, const uint8 w)
+void UCiFSocialNetwork::setWeight(const uint8 c1, const uint8 c2, const uint8 w)
 {
 	if (c1 < mNetwork.Num() && c2 < mNetwork.Num()) {
 		mNetwork[c1][c2] = w;
